got delete to work on container instead, handleClick magically worked again..
